chore: fix circular dependencies
