+++ packages/core/_src/stm/STM/operations/commit.ts
@@ -1,8 +1,1 @@
-/**
- * Commits this transaction atomically.
- *
- * @tsplus getter effect/core/stm/STM commit
- */
-export function commit<R, E, A>(self: STM<R, E, A>): Effect<R, E, A> {
-  return STM.atomically(self)
-}
+export { commit } from "@effect/core/stm/STM/definition/primitives"
