+++ packages/core/_src/stm/STM/operations/succeed.ts
@@ -1,10 +1,1 @@
-import { STMSucceedNow } from "@effect/core/stm/STM/definition/primitives"
-
-/**
- * Returns an `STM` effect that succeeds with the specified value.
- *
- * @tsplus static effect/core/stm/STM.Ops succeed
- */
-export function succeed<A>(a: A): STM<never, never, A> {
-  return new STMSucceedNow(a)
-}
+export { succeed } from "@effect/core/stm/STM/definition/primitives"
