+++ packages/core/_src/stm/STM/operations/sync.ts
@@ -1,10 +1,1 @@
-import { STMSucceed } from "@effect/core/stm/STM/definition/primitives"
-
-/**
- * Returns an `STM` effect that succeeds with the specified value.
- *
- * @tsplus static effect/core/stm/STM.Ops sync
- */
-export function sync<A>(a: LazyArg<A>): STM<never, never, A> {
-  return new STMSucceed(a)
-}
+export { sync } from "@effect/core/stm/STM/definition/primitives"
