+++ packages/core/_src/stm/TRef/operations/_internal/getOrMakeEntry.ts
@@ -1,5 +1,5 @@
+import type { Journal } from "@effect/core/stm/STM/definition/primitives"
 import { Entry } from "@effect/core/stm/STM/Entry"
-import type { Journal } from "@effect/core/stm/STM/Journal"
 
 export function getOrMakeEntry<A>(self: TRef<A>, journal: Journal): Entry {
   if (journal.has(self)) {
